Silently execute dependency monitoring tools (unless they fail)
Before, we'd dump the output to /dev/null.  Unfortunately, when things go
wrong the user gets no feedback.  Now, we monitor the process, and if it fails
dumps stdout and stderr.

Also a minor fix to the Python dependency monitor: a regexp that can handle
packages with '-' in the name.